pyproject.toml
[tool.poetry]
name = "slack-randompicker"
version = "0.1.0"
description = ""
authors = ["m-vdb"]

[tool.poetry.dependencies]
python = "^3.7"
pytest = "^5.4.1"
slackclient = "^2.5.0"
requests = "^2.23.0"
dateparser = "^0.7.4"
recurrent = "^0.2.5"
python-dateutil = "^2.8.1"
pytest-freezegun = "^0.4.1"
black = "^19.10b0"
mypy = "^0.770"
sanic = "^19.12.2"
apscheduler = "^3.6.3"
SQLAlchemy = "^1.3.16"
psycopg2 = "^2.8.5"
cron-descriptor = "^1.2.24"
pytest-asyncio = "^0.11.0"
pytest-mock = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
